gosdoc-backend/requirements/base.txt
# ============================================================
# ГосДок — базовые зависимости (раздел 7 ТЗ)
# ============================================================

# Django core
Django==5.1.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
django-filter==24.*

# Database
psycopg2-binary==2.9.*
dj-database-url==2.2.*

# Cache / Queue
celery==5.4.*
redis==5.0.*
django-redis==5.4.*
django-celery-beat==2.7.*

# AWS S3 / Object Storage
boto3==1.35.*
django-storages==1.14.*

# Document processing
PyMuPDF==1.24.*
python-docx==1.1.*

# AI
openai==1.40.*

# Report generation
weasyprint==62.*
openpyxl==3.1.*

# Security / Rate limiting
django-ratelimit==4.1.*
django-cors-headers==4.4.*

# Monitoring
sentry-sdk==2.13.*

# WSGI server
gunicorn==22.*

# API schema
drf-spectacular==0.27.*

# Utilities
python-decouple==3.8
Pillow==10.4.*
